requirements.txt
bcrypt==4.0.1
fastapi[all]==0.103.2
pandas==2.1.1
psycopg2-binary==2.9.8
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
uvicorn==0.23.2